+++ src/exports/standalone.index.ts
@@ -5,4 +5,5 @@ export { SnClippy } from '../components/sn-ui/sn-attachments/sn-clippy'
 export { SnActivity } from '../components/sn-ui/sn-activity/sn-activity'
 
 //Types
+export type { SnClippyRef } from '../components/sn-ui/sn-attachments/sn-clippy'
 export type { SnRecordPickerItem, SnRecordPickerList } from '../types/form-schema'